backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.0
tzdata==2025.2
celery==5.4.0
redis==5.0.7
drf-nested-routers==0.93.4
requests==2.32.3
Pillow==10.4.0 